+++ src/cjs/PlotBoilerplate.d.ts
@@ -90,8 +90,9 @@
  * @modified 2025-05-07 Handling content changes now with `contentChangeListeners`.
  * @modified 2025-05-07 Added `PlogBoilerplate.addContentChangeListener` and `.removeContentChangeListener`.
  * @modified 2025-05-07 Moving full vectors now by default when vector point a is moved.
+ * @modified 2025-05-20 Applying `lineWith` parameter in the draw routine for vectors (had been missing).
  *
- * @version  1.21.0
+ * @version  1.21.1
  *
  * @file PlotBoilerplate
  * @fileoverview The main class.