+++ src/cjs/PlotBoilerplate.js
@@ -91,8 +91,9 @@
  * @modified 2025-05-07 Handling content changes now with `contentChangeListeners`.
  * @modified 2025-05-07 Added `PlogBoilerplate.addContentChangeListener` and `.removeContentChangeListener`.
  * @modified 2025-05-07 Moving full vectors now by default when vector point a is moved.
+ * @modified 2025-05-20 Applying `lineWith` parameter in the draw routine for vectors (had been missing).
  *
- * @version  1.21.0
+ * @version  1.21.1
  *
  * @file PlotBoilerplate
  * @fileoverview The main class.
@@ -1220,7 +1221,7 @@ var PlotBoilerplate = /** @class */ (function () {
                 d.b.attr.renderTime = renderTime;
         }
         else if (d instanceof Vector_1.Vector) {
-            draw.arrow(d.a, d.b, this.drawConfig.vector.color);
+            draw.arrow(d.a, d.b, this.drawConfig.vector.color, this.drawConfig.vector.lineWidth);
             if (this.drawConfig.drawHandlePoints && d.b.attr.selectable && d.b.attr.visible) {
                 draw.setCurrentId("".concat(d.uid, "_h0"));
                 draw.setCurrentClassName("".concat(d.className, "-handle"));