+++ src/esm/utils/helpers/BezierPathInteractionHelper.js
@@ -26,6 +26,8 @@
  * @modified 2025-04-14 Fixing correct event types for touch events in `BezierPathInteractionHelper`.
  * @modified 2025-04-14 BezierPathInteractionHelper: Changed default value of `HelperOptions.autoAdjustPaths` from `true` to `false`.
  * @modified 2025-05-05 Added optional params `draw` and `fill` to BezierPathInteractionHelper.drawHandleLines` method.
+ * @modified 2025-05-05 Class `BezierPathInteractionHelper` now implementing `IShapeInteractionHelper`.
+ * @modified 2025-05-07 Tweaking performance of `BezierPathInteractionHelper`: only triggering redraw now when mouse move within given detect range.
  * @version  1.2.1
  *
  * @file BezierPathInteractionHelper
@@ -347,6 +349,9 @@ export class BezierPathInteractionHelper {
     // | Touch and mouse events should call this fuction when the pointer was moved.
     // +-------------------------------
     _handleMoveEvent(posX, posY) {
+        const oldA = this.currentA.clone();
+        const oldB = this.currentB.clone();
+        const oldDistance = this.currentDistance;
         const point = this.pb.transformMousePosition(posX, posY);
         this.currentB.set(point);
         this._updateMinDistance();
@@ -357,8 +362,12 @@ export class BezierPathInteractionHelper {
         else {
             this.onPointerMoved(-1, null, null, 0.0);
         }
-        // Always redraw even when moving outside the detection distance?
-        this.pb.redraw();
+        // Only redraw when moving inside the detection distance
+        if ((this.currentDistance <= this.maxDetectDistance ||
+            (this.currentDistance > this.maxDetectDistance && oldDistance <= this.maxDetectDistance)) &&
+            (!oldA.equals(this.currentA) || !oldB.equals(this.currentB))) {
+            this.pb.redraw();
+        }
     }
     // +---------------------------------------------------------------------------------
     // | This is called when the mouse pointer leaves the canvas or
@@ -374,23 +383,15 @@ export class BezierPathInteractionHelper {
     _installTouchListener() {
         var _self = this;
         const afProps = {
-            // Todo: which event types does AlloyFinger use?
-            // touchStart: function (_event: TouchEvent) {
-            //   _self.mouseIsOver = true;
-            // },
-            // /* TouchMoveEvent , AFTouchEvent<"touchStart"> */
             touchStart: function (_event) {
                 _self.mouseIsOver = true;
             },
             touchMove: function (event) {
-                // TouchEvent) {
                 if (_self.pb.getDraggedElementCount() == 0 && event.touches.length > 0) {
-                    // console.log('touchmove');
                     _self._handleMoveEvent(event.touches[0].clientX, event.touches[0].clientY);
                 }
             },
             touchEnd: function (_event) {
-                // TouchEvent) {
                 _self.mouseIsOver = false;
                 _self._clearMoveEvent();
             }
@@ -477,8 +478,9 @@ export class BezierPathInteractionHelper {
     // | Update the min distance from point `line.b` to the curve. And redraw.
     // +-------------------------------
     _updateMinDistance() {
-        if (this.paths.length == 0)
+        if (this.paths.length == 0) {
             return;
+        }
         let pathIndex = -1;
         let minDist = Number.MAX_VALUE;
         let closestPoint = null;